cpufeatures: minor code refactor
