+++ cpufeatures/src/lib.rs
@@ -5,30 +5,32 @@
     html_favicon_url = "https://raw.githubusercontent.com/RustCrypto/media/6ee8e381/logo.svg"
 )]
 
-#[cfg(not(miri))]
-#[cfg(target_arch = "aarch64")]
-#[doc(hidden)]
-pub mod aarch64;
-
-#[cfg(not(miri))]
-#[cfg(target_arch = "loongarch64")]
-#[doc(hidden)]
-pub mod loongarch64;
-
-#[cfg(not(miri))]
-#[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
-mod x86;
-
-#[cfg(miri)]
-mod miri;
-
-#[cfg(not(any(
-    target_arch = "aarch64",
-    target_arch = "loongarch64",
-    target_arch = "x86",
-    target_arch = "x86_64"
-)))]
-compile_error!("This crate works only on `aarch64`, `loongarch64`, `x86`, and `x86-64` targets.");
+cfg_if::cfg_if!(
+    if #[cfg(miri)] {
+        mod fallback;
+    } else if #[cfg(all(target_arch = "aarch64", any(target_os = "linux", target_os = "android")))] {
+        #[doc(hidden)]
+        pub mod aarch64_linux;
+    } else if #[cfg(all(target_arch = "aarch64", target_vendor = "apple"))] {
+        #[doc(hidden)]
+        pub mod aarch64_apple;
+    } else if #[cfg(all(target_arch = "loongarch64", target_os = "linux"))] {
+        #[doc(hidden)]
+        pub mod loongarch64_linux;
+    } else if #[cfg(any(target_arch = "x86", target_arch = "x86_64"))] {
+        // CPUID is not available on SGX. Freestanding and UEFI targets
+        // do not support SIMD features with default compilation flags.
+        cfg_if::cfg_if!(
+            if #[cfg(any(target_env = "sgx", target_os = "none", target_os = "uefi"))] {
+                mod fallback;
+            } else {
+                mod x86;
+            }
+        );
+    } else {
+        mod fallback;
+    }
+);
 
 /// Create module with CPU feature detection code.
 #[macro_export]
@@ -37,7 +39,7 @@ macro_rules! new {
         mod $mod_name {
             use core::sync::atomic::{AtomicU8, Ordering::Relaxed};
 
-            const UNINIT: u8 = u8::max_value();
+            const UNINIT: u8 = u8::MAX;
             static STORAGE: AtomicU8 = AtomicU8::new(UNINIT);
 
             /// Initialization token
@@ -48,10 +50,12 @@ macro_rules! new {
                 /// Get initialized value
                 #[inline(always)]
                 pub fn get(&self) -> bool {
-                    $crate::__unless_target_features! {
-                        $($tf),+ => {
-                            STORAGE.load(Relaxed) == 1
-                        }
+                    if cfg!(all($(target_feature=$tf,)*)) {
+                        true
+                    } else if $crate::__can_detect!($($tf),+) {
+                        STORAGE.load(Relaxed) == 1
+                    } else {
+                        false
                     }
                 }
             }
@@ -60,24 +64,26 @@ macro_rules! new {
             /// initializing underlying storage if needed.
             #[inline]
             pub fn init_get() -> (InitToken, bool) {
-                let res = $crate::__unless_target_features! {
-                    $($tf),+ => {
-                        #[cold]
-                        fn init_inner() -> bool {
-                            let res = $crate::__detect_target_features!($($tf),+);
-                            STORAGE.store(res as u8, Relaxed);
-                            res
-                        }
-
-                        // Relaxed ordering is fine, as we only have a single atomic variable.
-                        let val = STORAGE.load(Relaxed);
-
-                        if val == UNINIT {
-                            init_inner()
-                        } else {
-                            val == 1
-                        }
+                let res = if cfg!(all($(target_feature=$tf,)*)) {
+                    true
+                } else if $crate::__can_detect!($($tf),+) {
+                    #[cold]
+                    fn init_inner() -> bool {
+                        let res = $crate::__detect!($($tf),+);
+                        STORAGE.store(res as u8, Relaxed);
+                        res
                     }
+
+                    // Relaxed ordering is fine, as we only have a single atomic variable.
+                    let storage_val = STORAGE.load(Relaxed);
+
+                    if storage_val == UNINIT {
+                        init_inner()
+                    } else {
+                        storage_val == 1
+                    }
+                } else {
+                    false
                 };
 
                 (InitToken(()), res)
